Reorganised gdl modules, updated setup.py, other fixes
