+++ .gitignore
@@ -7,3 +7,4 @@ installed-files.txt
 /pysilfont.egg-info
 /setuptools-*
 /local
+pysilfont.cfg
