+++ setup.py
@@ -28,7 +28,12 @@ setup(
 		maintainer='NRSI - SIL International',
 		maintainer_email='fonts@sil.org',
 		url='http://github.com/silnrsi/pysilfont',
-		packages = ["silfont"],
+		packages = ["silfont",
+            "silfont.compdef",
+            "silfont.ftml",
+            "silfont.gdl",
+            "silfont.ufo",
+                   ],
 		package_dir = {'':'lib'},
 		scripts=['scripts/UFOconvert', 'scripts/makeGdl'],
 		license='MIT',
